refactor Row class into its own file
